[refactor] moving main class
